Update getEdgesDirections queries
Replaces the queries to fetch edge directions while building the schema.
The old queries worked by testing for the existance of any possible edge,
which resulted in a large number of queries for graphs with lots of node and
edge labels.

The new query instead iterates through all the edge labels, and runs a query
to fetch the corresponding from-label to-label pairs.