Refactor console output and introduce file log levels (#6)
Refactored console output and spinners to be more like a command line progress UI and introduced debug level file logs. Setting the verbose-log option to true will log more data to the log file. Since the verbose log data can include graph data values it is intended for the user to have to opt in to this verbose logging. Changed pino standard out logger to console as the pino logger was causing interleaving log messages that were out of order with the spinners. Ensured to not log to console if a spinner is spinning as this can have unintended interleaving of output.