Added separate input types for create and update mutations while maintaining original input type (#93)
The utility currently generates the same input types for queries as mutations.

This is not ideal as it would be expected that the required fields will differ for queries vs mutations. For example, id should not be mandatory when creating a node as it should be auto generated by the server. However id should be required when updating a node.

Generating distinct input types for queries vs create and update mutations will provide better flexibility to the user to modify the generated schema accordingly.