Added support for setting an offset value as part of query options. (#102)
Added support for setting an offset value as part of query options which can be used in combination with the existing limit for pagination purposes. The offset is translated to open cypher's SKIP by the resolver. If offset and limit are specified for nested edge queries, the values are translated to open cypher's list slice lower and upper bounds.