Replace graphql-tag.gql with graphql.parse for resolver (#110)
Replace use of graphql-tag gql with graphql parse for server resolver code as gql can resolve stale cached variable values if the same variable name is referenced in subsequent queries. According to graphql-tag documentation, the library is meant for client-side code so may not be suitable for server-side usage.

Note that this change fixes stale variable values on the resolver side but it is still possible that cached variable values can be referenced on the client-side. For example the default Apollo Studio UI and client for local environments have their own caching mechanisms and are out of scope for this changeset.

There is an additional change to update the brace-expansion version as the previous one had a minor vulnerability.