fix: Updated index.mjs file to be modified for Apollo subgraph Federation 2 features (#126)
There was an issue with the Apollo index.mjs file when generated with --create-update-apollo-server-subgraph as it did not include support for the Apollo Subgraph Federation 2 features.
Updated so that when user runs --create-update-apollo-server-subgraph the index.mjs file includes the support for the subgraph federation 2 features and if run with just --create-update-apollo-server, the index.mjs file is not modified and same as the original template.
Also added an index.mjs file check to the apollo subgraph integration test (Case09) to ensure the index.mjs file is always modified correctly to add support for the subgraph federation 2 features.