bug: Accommodated for special characters in edge labels to ensure correct Cypher translation (#127)
Queries would fail if an edge label contained a special character such as a colon, period, or space as it is invalid formatting for Cypher. Thus, back ticks were added to ensure that special characters could be included as label names without crashing from incorrect Cypher translations.

Added a unit test for the resolver to ensure backticks are added correctly in cypher translations from queries with special characters in edge labels. Also changed existing resolver unit tests to accommodate for this change.