Added Apollo Server GraphQL API error message for failed requests (#129)
Any expired security tokens or network connectivity errors when querying neptune resulted in a cryptic error message that was not helpful for debugging.

Instead of swallowing errors, added an error catch message with rethrow to ensure that when these errors result it returns a more useful error message that will help with the debugging of the error.